flattening file structure
