+++ lib/AnalyzeTestResults.js
@@ -3,7 +3,7 @@ const path = require('path');
 const { XMLParser } = require('fast-xml-parser');
 
 const getCliArgs = require('../GetCliArgs/GetCliArgs');
-const getProjectPath = require('../GetProjectPath/GetProjectPath');
+const getProjectPath = require('./GetProjectPath');
 
 function analyzeTestResults() {
     const args = getCliArgs();
