+++ lib/DownloadGodot.js
@@ -1,17 +1,81 @@
 const https = require('https');
 const path = require('path');
 const fs = require('fs');
-
 const decompress = require('decompress');
-
-const {
-    createTmpDir,
-    deleteTmpDir,
-    getTmpDir
-} = require('./helpers/TmpDirHelpers/TmpDirHelpers');
-const performDownload = require('./helpers/PerformDownload/PerformDownload');
-const generateGodotFilename = require('./helpers/GenerateGodotFilename/GenerateGodotFilename');
 const getCliArgs = require('../GetCliArgs/GetCliArgs');
+
+function getTmpDir() {
+    if(!process.env.THIS_ACTION_DIR) {
+        throw new Error('THIS_ACTION_DIR environment variable not set');
+    }
+
+    return path.join(process.env.THIS_ACTION_DIR, 'tmp');
+}
+
+function deleteTmpDir() {
+    const tmpDir = getTmpDir();
+    if (fs.existsSync(tmpDir)) {
+        fs.rmdirSync(tmpDir, { recursive: true });
+    }
+};
+
+function createTmpDir() {
+    const tmpDir = getTmpDir();
+    if (!fs.existsSync(tmpDir)) {
+        fs.mkdirSync(tmpDir);
+    }
+};
+
+function performDownload(url, fileLocation) {
+    const downloadedFile = fs.createWriteStream(fileLocation);
+
+    return new Promise((resolve, reject) => {
+        https.get(url, response => {
+            response.pipe(downloadedFile);
+            downloadedFile.on('finish', () => {
+                downloadedFile.close();
+                resolve(fileLocation);
+            });
+        }).on('error', error => {
+            fs.unlink(fileLocation);
+            reject(error);
+        });
+    });
+}
+
+function generateGodotFilename() {
+    const args = getCliArgs();
+
+    const godotVersion = args.godotVersion;
+    const releaseType = args.releaseType;
+    const isMono = args.isMono;
+
+    let fullGodotName = `Godot_v${godotVersion}-${releaseType}`;
+    let fullGodotNameWithArch = '';
+    if (isGodotFour()) {
+        if (isMono) {
+            fullGodotName += '_mono_linux';
+            fullGodotNameWithArch = `${fullGodotName}_x86_64`;
+        } else {
+            fullGodotName += '_linux';
+            fullGodotNameWithArch = `${fullGodotName}.x86_64`;
+        }
+    } else {
+        if (isMono) {
+            fullGodotName += '_mono_linux_headless';
+            fullGodotNameWithArch = `${fullGodotName}_64`;
+        } else {
+            fullGodotName += '_linux_headless';
+            fullGodotNameWithArch = `${fullGodotName}.64`;
+        }
+    }
+
+    return {
+        fullGodotName,
+        fullGodotNameWithArch
+    };
+};
+
 
 function generateDownloadUrl() {
     const {
