+++ lib/ExecuteGutTests.js
@@ -1,6 +1,6 @@
-const getProjectPath = require('../GetProjectPath/GetProjectPath');
+const getProjectPath = require('./GetProjectPath');
 const getCliArgs = require('../GetCliArgs/GetCliArgs');
-const isGodotFour = require('../IsGodotFour/IsGodotFour');
+const isGodotFour = require('./IsGodotFour');
 const { spawnSync } = require('child_process');
 
 function executeGutTests(exePath) {
