+++ lib/GetProjectPath.js
@@ -1,5 +1,5 @@
 const pathLib = require('path');
-const getCliArgs = require('../GetCliArgs/GetCliArgs');
+const getCliArgs = require('./GetCliArgs');
 
 function getProjectPath() {
     if (!process.env.GITHUB_WORKSPACE) {
