+++ lib/IsGodotFour.js
@@ -1,4 +1,4 @@
-const getCliArgs = require('../GetCliArgs/GetCliArgs');
+const getCliArgs = require('./GetCliArgs');
 
 function isGodotFour() {
     const args = getCliArgs();
