+++ lib/RunGodotImport.js
@@ -1,6 +1,8 @@
-const fs = require('fs');
-const getProjectPath = require('../../../GetProjectPath/GetProjectPath');
 const path = require('path');
+const { spawnSync } = require('child_process');
+
+const getProjectPath = require('./GetProjectPath');
+const getCliArgs = require('./GetCliArgs');
 
 function addRebuilderSceneToProjectAddonFolder() {
     if(!process.env.THIS_ACTION_DIR) {
@@ -62,7 +64,32 @@ function removeRebuilderSceneFromProjectAddonFolder() {
     });
 };
 
-module.exports = {
-    addRebuilderSceneToProjectAddonFolder,
-    removeRebuilderSceneFromProjectAddonFolder,
+function runGodotImport(exePath) {
+    addRebuilderSceneToProjectAddonFolder();
+
+    console.log("Running Godot import...");
+    const { importTime } = getCliArgs();
+    const importProcResults = spawnSync(
+        exePath,
+        [
+            '--headless',
+            '--editor',
+            path.join(
+                'addons',
+                'gut',
+                '.cli_support',
+                '__rebuilder_scene.tscn',
+            ),
+        ],
+        {
+            cwd: getProjectPath(),
+            timeout: importTime * 1000,
+        }
+    );
+
+    console.log('Godot Import results: ', importProcResults.stdout.toString());
+
+    removeRebuilderSceneFromProjectAddonFolder();
 };
+
+module.exports = runGodotImport;
