+++ test/IsGodotFour.test.js
